Clean up src/ and tests/ directories: Remove obsolete files and fix dependencies
Frontend (src/) cleanup:
- Fix App.jsx: Remove imports for deleted audioCodecMonitor and memoryPressureMonitor
- Delete src/utils/stabilityMonitor.test.js (unused test file)
- Update vite.config.js: Remove references to deleted monitoring utilities

Tests cleanup:
- Convert CommonJS test files from .js to .cjs extension for compatibility
- Remove duplicate test files (.js versions, keep .cjs)
- Update run-tests.cjs to scan for .cjs test files
- All tests passing: 11/11 successful

Preserved essential files:
- All actively used services, hooks, components, and utilities
- All production-critical functionality maintained
- Build and test systems fully functional

Build status: ✅ Successful
Test status: ✅ All 11 tests passing
