+++ tests/server-health.test.cjs
@@ -16,24 +16,24 @@ process.env.GEMINI_TTS_ENDPOINT = 'https://gemini.test/tts';
 
 const app = require('../backend/server');
 
-function assert(cond, msg){ if(!cond) throw new Error(msg || 'Assertion failed'); }
+function assert(cond, msg) { if (!cond) throw new Error(msg || 'Assertion failed'); }
 
 exports.health_ok = () => new Promise((resolve, reject) => {
-  const server = app.listen(0, () => {
-    const port = server.address().port;
-    http.get({ hostname: '127.0.0.1', port, path: '/health', timeout: 2000 }, res => {
-      let data='';
-      res.on('data', d=>data+=d);
-      res.on('end', () => {
-        try {
-          assert(res.statusCode === 200 || res.statusCode === 503, 'Status 200 veya 503 olmalı');
-          const json = JSON.parse(data);
-          assert(json.status, 'status alanı olmalı');
-          assert(json.services && json.services.database, 'services.database olmalı');
-          server.close();
-          resolve();
-        } catch(e){ server.close(); reject(e); }
-      });
-    }).on('error', err => { server.close(); reject(err); });
-  });
+    const server = app.listen(0, () => {
+        const port = server.address().port;
+        http.get({ hostname: '127.0.0.1', port, path: '/health', timeout: 2000 }, res => {
+            let data = '';
+            res.on('data', d => data += d);
+            res.on('end', () => {
+                try {
+                    assert(res.statusCode === 200 || res.statusCode === 503, 'Status 200 veya 503 olmalı');
+                    const json = JSON.parse(data);
+                    assert(json.status, 'status alanı olmalı');
+                    assert(json.services && json.services.database, 'services.database olmalı');
+                    server.close();
+                    resolve();
+                } catch (e) { server.close(); reject(e); }
+            });
+        }).on('error', err => { server.close(); reject(err); });
+    });
 });
